Add experimental sticky titles theme
Select the new "Sticky Titles" theme to cause tiddler titles to stick
to the top of the window as you scroll. It works very well in Chrome
Canary, reasonably well in iOS 7.

As an aside, it would be nice to be able to package this is as
something a bit smaller than a theme; maybe themes need to have
optional hacks that can be associated with them.