Update the Snow White theme to use px for dimensions instead of em
The old behaviour meant that resizing the base font resized every page
element in proportion. The new behaviour makes the text resize within a
fixed layout, so that one can alter the number of characters displayed
on each line of a tiddler. The old behaviour can usually be emulated by
using the browser page zoom tool.