First pass at refactored widget implementation
Now there's a single widget class used for the render tree, with
subclasses for widget instances. Text nodes and elements are now
represented as widgets.

It is now possible for a widget not to generate any DOM nodes (eg the
transclude widget).

This first commit has got the basics, now we need to add the rest of
the existing widgets, and then remove the old rendertree stuff.