Remove text input showing static URL preview
It wasn’t working due to the recent change whereby attributes set as
macros are no longer wikified. Thus, the widget invocation inside the
macro outerMakeLink isn’t executed.

To get it working again we’ll need a way to specify a macro as an
attribute and force it to be wikified.