Introducing lazy-edit
This introduces lazy edit of edit-text where the field will be updated on blur. This should fix (or workaround) many of the "input field looses focus" problems.

The optional attribute "edit" can be set to "lazy". In this case, the "input" event won't trigger an update, but the "blur" event will.