Add create param
Adds a create param to the fieldmangler widget. The create param must be a valid JSON string, and does not need to contain the title. The title will be set to the tiddler param regardless of whether it is specified in create. 

The docs line could read: | create | Create the specified tiddler using these fields if it doesn't exist. |