Remove the scrollable from the story river
The problem was that the scrollable required that the mouse be over the
story river itself in order to scroll it with the wheel. If we fallback
to scrolling the story by scrolling the browser body then we are able
to scroll even if the mouse is over the sidebar.