Add create param to fieldmangler.js
Desierializes the text in the create attribute. On failure, the desierializer returns a fields list containing only an empty text field, which is good. This only adds the tiddler if it doesn't already exist! Because this is the fieldmangler, the title of the created tiddler is automatically set to the mangleTitle. The tiddler is NOT opened at all.