Update handleNewTiddlerEvent().
Changes the handling of the param attribute if it contains a line break. The entire param string is deserielized as application/x-tiddler and the resulting tiddler is used as the template tiddler. This is as opposed to getting the tiddler from the store with only the title. This gets rid of the requirement for a separate template tiddler. This is mostly for hackability. Good plugins should have a separate template tiddler to ease upgrades.