Allow anonymous routes / allow query parameters
Hi Jeremy!
For setting up TW Server in an Amazon Cloud compatible environment I was required to have a loadbalancer in front of my zone. Unfortunately I could only use a "Ping" healthcheck because all routes require the user to be logged in. This change allows anonymouse routes.

A small additional change is the "true" in line 89 which will make url.parse also parse the query string which eases creation of routes needing them. I needed it for my "publish" route which allows to define templates like so: .../publish/<TiddlerTioPublish>.html?template=<TemplateTiddler>