Move the username for signing edits from a tiddler into browser local-storage.
Actually, this is only done in the common case; the old tiddler, though no longer exposed in the control panel, will still be used for situations where local storage isn't available, such as running server-side.

Also, refactor slightly, changing the DOM-util functions for managing passwords in local storage to be general to all types of data, changing everything that used them accordingly, and adding a widget for non-password values kept in the browser.

This should all be backward-compatible in the sense that users don't need to be aware anything has changed.  Specifically, the upload password is still kept in exactly the same place; usernames in the old place will continue to work; and the new place for usernames will be used the next time it is set in the control panel.