Refactor theme tweaks as individual tiddlers
The trouble with the previous approach of storing several settings in a
single data tiddler was that it made it hard to introduce new settings.
Users who already have a custom copy of the settings/metrics tiddler
wouldn’t get the default settings. The new approach uses individual
tiddlers for each setting.

This commit also introduces a new theme tweak for the cutoff width for
switching between responsive views.