Add direct link to github ribbon plugin
The former documentation referenced a plugin but left no easy way to
link to it. Also, the misleading text claiming a title also confused me.
I had to do some creative greping the git files to hunt it down.

Adding a TiddlyWiki link around the plugin links to the actual tiddler
that a reader could easily copy. This also deprecated the need for the
misleading title.