Proposal to allow creating a field and setting its value in one go
I always find it very tedious to first enter a field name, click "add" (could you add "Return" here as a shortcut?) and then to click the newly created field to enter the value. Especially when an instruction shows field and value like this: "field: value".

This proposal splits a field name at the first colon and sets the name to be the left and the new value to be the right part.