House style tweaks
@giffmex - I’m sorry I should document these things but you’ll see I’ve
made some changes to punctuation and whitespace to match other tw5.com
tiddlers.

I’ve just noticed that the final bullet point ended abruptly with
“Finally, you will want a tiddler”. I’ve removed that line for the
moment.

The section “Add or remove shadow tiddlers from the user interface” is
actually describing a technique that is not recommended. If you remove
a tag from a shadow tiddler then you’ll have made a copy of it, which
means that it won’t be upgraded in future. It’s still worth
documenting, but with some warnings.

I do wonder whether the three sections deserve to be separate tiddlers.
Certainly the instructions for adding a toolbar button should be
directly addressable so that we can link directly to them on
http://tiddlywiki.com.