edit-tags-filter-tags-and-titles
fixes #1333

extracts the tag-link into a component template...

**$:/core/ui/Components/tag-link**

...so as to split the list in two, having two filters that list tiddlers
in the add-tags popup to...

1. firstly, list matching used tags
2. secondly, list mathcing tiddler titles, thus availabe for tagging

Does away with searching content as it is irrelevant if not distracting
for tagging.

Why two lists? Consider wanting to tag a tiddler with "Filters"...
observe how the result would otherwise be way too far down because
existing tiddler titles get in the way.

Component templates are also desireable elsewhere, e.g. #1318