Extend link widget with "draggable", "tabindex" and "tag"
Allows us to disable draggability of links, specify a tabindex for
keyboard control, and specify the generated HTML tag.

This allows us to use a link as the tiddler frame in the story view,
which means that when you click on the background of a tiddler it is
navigated. In the stacked story view, that means that you can click on
a tiddler to bring it the front.

But it’s not perfect. For example, selecting text in classic view
causes the tiddler to scroll to the top afterwards.