Add first pass of help panel
This is just an idea really.

The help plugin is only included in the tw5.com wiki; the idea is to
also include it in the empty.html wiki, either as a plugin or as part
of the core.

If it stays as a plugin, the idea would be to have different help
plugins for different languages.