Turn tiddler frames into links
By turning the tiddler frame into a link it is now possible to click on
a tiddler to navigate to it. This is needed for the new
stacked-storyview, but it also useful in other storyviews.

Note that there are currently some issues. For example, in classic
storyview, clicking on a tiddler will scroll to the top of it, which
isn’t helpful if you’ve scrolled halfway down a long tiddler.