Make tiddler frames clickable
With this enhancement, clicking on a tiddler in the story view will
navigate to that tiddler. This is needed for eg the stacked storyview,
where we want to be able to click on tiddlers to bring them to the
front of the stack.

There are some problems though - as things stand, clicking on a tiddler
in classic storyview will scroll to the top of that tiddler.