Remove click-to-navigate on tiddler bodies
This was introduced to make it possible in the new stacked story view
to click on a tiddler to bring it to the front. By allowing a tiddler
in view mode to be focussed, it also prepared for view mode keyboard
shortcuts. (Eg, a key to edit the current tiddler).

However, there are several minor issues with the unexpected behaviour
of clicking on a tiddler navigating to it, so we’ll leave this out of
5.1.8