Revert to original fix for #1542
The trouble with tweaking the ViewBox was that the amount of the
adjustment is expressed in the coordinate system of the image, not the
coordinate system of the bitmap that is being rendered. That means that
the additional space doesn’t necessarily extend to the single physical
pixel needed to resolve the issue.