Remove cheatsheet from help panel
We’d rather present this information as a panel within the tiddler
editor so that we can (for example), allow clicks on a snippet to
insert it automatically into the editor at the current cursor position.

For the moment, we’ll just remove the cheatsheet. Post-5.1.8 we’ll add
it back to the editor