When using TiddlyWiki I often find myself in the situation, where I rather want to show the tiddler with a short title. For example when it is listed as tab or when it is displayed tag. In case of tabs, TiddlyWiki acknowledges the caption field. But in case of tags - the whole title is used. In most of the cases I see myself forced to use a very small title because I also want to use it as tag, which is not nice.
Therefore I propose the following change which takes advantage of the fact that "the content of the <$view> widget is displayed if the field or property is missing or empty":

For a tiddler that is *not* in draft mode:

1. Try to display the caption field as tag label
2. If the field does not exist or is empty, use the title

**This proposal does not break the reference-system**, as the correct title has to be entered to reference the tiddler.

I would also suggest to display the short title in the drop down. But want to discuss this first.