Button to hide the contents of plugin libraries
This just adds a button and reveal around the list of tiddlers contained in a plugin library so that they can be hidden if desired.

When you have multiple libraries open the wiki can get confused about which one you are looking at when you change between the plugin, theme and languages tabs. There is strange scrolling behaviour and it seems to give focus to the search box from one of the libraries chosen at random.

Also when you have multiple libraries open the list can get quite long and being able to hide them makes it look a bit cleaner.