Add ability to auto generate Dropbox URL
I felt the conversion from the Dropbox share URL to the "secret" URL was
complicated enough to have the tiddler do it for the user. And so I
wrote a widget to do that. The user enters in the shared URL given by
Dropbox and the widget outputs the secret URL that they can copy / paste
with.

See if it is worth having on the TW5.com site.