Added mechanism to rename tags
TiddlyWiki's strength is to allow a user to manipulate or reorganize content/data/meta-data in the wiki as fast as the person's mind changes thoughts. This way the wiki is always an explicit model of the current mind state.

The tag mechanism is an important part here but it doesn't allow renaming tags, which is a major drawback: If a user created a tag and in the course of building the wiki discovers that the tag is not appropriately named, he/she has to edit maybe 30 tiddlers to rename a single tag.

Naturally, many users asked about a feauture to rename tags.

* https://groups.google.com/forum/#!searchin/tiddlywiki/rename$20tag|sort:date/tiddlywiki/P_el-4lg_P0/5WgyPcTW8I8J
* https://groups.google.com/forum/#!searchin/tiddlywiki/rename$20tag|sort:date/tiddlywiki/uuNeYfa5d0U/txJAIoU8aA4J
* https://groups.google.com/forum/#!searchin/tiddlywiki/rename$20tag/tiddlywiki/5vjg5EXSXrw/Dp8wSz4H6gwJ
*...

In response to #1614 I now did the PR you are looking at.

See a demo at http://wkpr.de/hosting/tmp/tw5/tagrenaming/

I think the GUI can be optimized but this is just a demo.

How it works

1. Tags will be renamed in all tiddlers containing the tag
1. Finally the tag tiddler (if exists) will be renamed

Jed's wikitext based method is extremy smart(!) but it is not javascript based which prevents it from being invoked from js-code.

http://inmysocks.tiddlyspot.com/#Draft%20of%20%27Search%20and%20Replace%20Tags%27

The PR doesn't contain lingo-stuff yet. This will be added at the end after discussion.