My first attempt at editing the TW documentation
Feel free to ignore this pull request if you want to, although it might be a useful addition to this page. I use NZ English normally, which is closer to UK English than to US English, but I still get confused between the two sometimes, having spent so long around the (predominately American) Internet.