Fix a bug whereby the text field of the current tiddler may be inadvertently cleared
If the `$field` and `$value` attributes are not set, the 'text' field of the current tiddler gets cleared (default field="text", default value="")

Whilst I was at it -- I added the facility to set any number of TextReferences. This means that values may be set for more than one data index, and different tiddlers may be targeted with the same widget.