Add documentation for negation
Also, I don't quite like the term "Filter Parameter" as meaning "Filter" or "Filter Operand". The truth is, **all** of prefix, suffix and operand are **filter parameters**, not just the operand! Also, in terms of wording, a step specifies a **filter** via the operator, but a filter step is **not** the same as a given filter (operator).