Modified CheckboxWidget and docs
After completion of the Checklist widget, I have here utilised some of
the code to extend the functionality of the Checkbox widget to handle
the toggling of a tag or string value in an index.

The widget will not write to an index unless the target type has been
set to ‘Data-Dictionary’ — this guards against overwriting the ‘text’
field of the target unintentionally.