Add "new tab" button to sidebar
Adding a tab to the sidebar is a common operation when using TiddlyWiki
for organizing notes. Currenty, new users have to find the relevant
howto in the documentation, enter the correct tag and copy-paste the
macro call. Adding a button after the regular sidebar tabs makes all
of this easily discoverable, and streamlines the process also for
experienced users.

The button can be switched off in the Settings tab of the $:/ControlPanel.